refactor: remove space modal directory
